4's commit
Discription:
- finish project

+++ ylab_hw/src/api/v1/resources/posts.py
@@ -1,10 +1,13 @@
 from http import HTTPStatus
 from typing import Optional
+
 from fastapi import APIRouter, Depends, HTTPException
+
 from src.api.v1.schemas import PostCreate, PostListResponse, PostModel
-from src.services.post import PostService, get_post_service
 from src.models.users import User
+from src.services.post import PostService, get_post_service
 from src.services.user_services import UserService, get_user_service
+
 from .auth import AuthHandler
 
 
@@ -26,6 +29,7 @@ def post_list(
         # Если посты не найдены, отдаём 404 статус
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="Posts not found.")
     return PostListResponse(**posts)
+
 
 @router.get(
     path="/{post_id}",
@@ -53,13 +57,13 @@ def patch_post_detail(
     payload_info: User = Depends(auth_handler.auth_current_user_uuid),
     user_checker: UserService = Depends(get_user_service),
 ) -> PostCreate:
-    if user_checker.check_access_tkn(payload_info[0]['jti']):
+    if user_checker.check_access_tkn(payload_info[0]['jti']) and payload_info[0]['type'] == "access":
         patch_post: Optional[dict] = post_service.update_post_detail(post_id, post)
         if not patch_post:
             raise HTTPException(HTTPStatus.NOT_FOUND, detail="Post not found.")
         return {"msg": "Post updated.", "Updated post": patch_post}
     else:
-        raise HTTPException(status_code=401,
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED,
                             detail="Could not validate user. Login once again or refresh access_token.")
 
 
@@ -75,11 +79,11 @@ def post_create(
     payload_info: User = Depends(auth_handler.auth_current_user_uuid),
     user_checker: UserService = Depends(get_user_service),
 ) -> PostCreate:
-    if user_checker.check_access_tkn(payload_info[0]['jti']):
+    if user_checker.check_access_tkn(payload_info[0]['jti']) and payload_info[0]['type'] == "access":
         post: dict = post_service.create_post(post=post)
         return PostModel(**post)
     else:
-        raise HTTPException(status_code=401,
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED,
                             detail="Could not validate user. Login once again or refresh access_token.")
 
 
@@ -88,17 +92,16 @@ def post_create(
     summary="Удалить определенный пост",
     tags=["posts"],
 )
-def post_detail(
+def post_delete(
     post_id: int, post_service: PostService = Depends(get_post_service),
     payload_info: User = Depends(auth_handler.auth_current_user_uuid),
     user_checker: UserService = Depends(get_user_service),
 ) -> PostCreate:
-    if user_checker.check_access_tkn(payload_info[0]['jti']):
+    if user_checker.check_access_tkn(payload_info[0]['jti']) and payload_info[0]['type'] == "access":
         result = post_service.delete_post(post_id=post_id)
         if not result:
             raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="Post not found.")
         return result
     else:
-        raise HTTPException(status_code=401,
+        raise HTTPException(status_code=HTTPStatus.UNAUTHORIZED,
                             detail="Could not validate user. Login once again or refresh access_token.")
-
